Rename puzzle history screen file
